+++ test/field.telephone.spec.js
@@ -36,14 +36,7 @@ describe('telephone',  () => {
 
   });
 
-  it('leave empty field', async () => {
-    let f2 = new FieldTelephone({ lookup: new Lookup(), removeEmpty : false});
-    let r = await f2.convert('telephone', {telephone: ''}, logger);
-    assert.isDefined(r.value, 'still there');
-    assert.equal(r.value, '', 'and empty')
-  });
-
-  it('process', async () => {
+ it('process', async () => {
     class TelLook extends Lookup {
     }
     let f2 = new FieldTelephone({ lookup: new TelLook(), removeEmpty : false});
@@ -94,5 +87,13 @@ describe('telephone',  () => {
       assert.isDefined(r._mode);
       assert.equal(r._mode, 1)
     });
+  });
+
+  describe('empty telephone', async() => {
+    let f = new FieldTelephone();
+    it('empty', async () => {
+      let r = await f.convert('telephone', {type: '1234', _mode: 'add'}, logger);
+      assert.equal(Object.keys(r).length, 0, 'has to remove everything');
+    })
   })
 });